requirements.txt
requests
beautifulsoup4
psycopg2-binary
pandas
python-dotenv
openai
slack-sdk
streamlit
plotly
pytest
apache-airflow
dbt-postgres
